+++ src/background/seelen_wm_v2/hook.rs
@@ -1,9 +1,18 @@
+use lazy_static::lazy_static;
+use parking_lot::Mutex;
+use seelen_core::rect::Rect;
+use std::sync::Arc;
+
 use crate::{
-    error_handler::Result, modules::virtual_desk::VirtualDesktopEvent, windows_api::window::Window,
-    winevent::WinEvent,
+    error_handler::Result, modules::virtual_desk::VirtualDesktopEvent, trace_lock,
+    windows_api::window::Window, winevent::WinEvent,
 };
 
-use super::WindowManagerV2;
+use super::{cli::Axis, state::WM_STATE, WindowManagerV2};
+
+lazy_static! {
+    static ref SystemMoveSizeStartRect: Arc<Mutex<Rect>> = Arc::new(Mutex::new(Rect::default()));
+}
 
 impl WindowManagerV2 {
     pub fn process_vd_event(event: &VirtualDesktopEvent) -> Result<()> {
@@ -30,10 +39,46 @@ impl WindowManagerV2 {
             WinEvent::SystemMoveSizeStart => {
                 if Self::is_managed(window) {
                     Self::set_overlay_visibility(false)?;
+                    *SystemMoveSizeStartRect.lock() = window.inner_rect()?;
                 }
             }
             WinEvent::SystemMoveSizeEnd => {
                 if Self::is_managed(window) {
+                    let initial_rect = SystemMoveSizeStartRect.lock();
+                    let initial_width = (initial_rect.right - initial_rect.left) as f32;
+                    let initial_height = (initial_rect.bottom - initial_rect.top) as f32;
+
+                    let rect = window.inner_rect()?;
+                    let new_width = (rect.right - rect.left) as f32;
+                    let new_height = (rect.bottom - rect.top) as f32;
+
+                    if initial_width != new_width {
+                        let percentage_diff = (new_width - initial_width) / initial_width * 100.0;
+                        let axis = if rect.left == initial_rect.left {
+                            Axis::Right
+                        } else {
+                            Axis::Left
+                        };
+                        log::trace!("window width changed by: {}%", percentage_diff);
+                        let state = trace_lock!(WM_STATE);
+                        let (m, w) = state.update_size(window, axis, percentage_diff, true)?;
+                        Self::render_workspace(&m.id, w)?;
+                    }
+
+                    if initial_height != new_height {
+                        let percentage_diff =
+                            (new_height - initial_height) / initial_height * 100.0;
+                        let axis = if rect.top == initial_rect.top {
+                            Axis::Bottom
+                        } else {
+                            Axis::Top
+                        };
+                        log::trace!("window height changed by: {}%", percentage_diff);
+                        let state = trace_lock!(WM_STATE);
+                        let (m, w) = state.update_size(window, axis, percentage_diff, true)?;
+                        Self::render_workspace(&m.id, w)?;
+                    }
+
                     Self::force_retiling()?;
                     Self::set_overlay_visibility(true)?;
                 }