+++ changelog.md
@@ -24,6 +24,7 @@
 - allow drop files, apps and folders into the dock to pin them.
 - now Virtual Desktop shortcuts doesn't require Tiling WM be enabled to work.
 - now Themes are wrapped in a CSS layer, making easier the override theming.
+- allow change size of Window Manager Layouts via window resizing with the mouse.
 
 ### fix
 - missing icons for files with a different extension than `exe`.